+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### ADD
+- **`feature` `#03 - Mockito` // `#043` - Argument matchers**
+  - _The ExamServiceImplTestData class has been updated._
+  - _A new test has been added to the `ExamServiceImplTest` class to test some "Argument Matchers"._  
+  - _The NOTES file has been updated._  
+  
+
 - **`feature` `#03 - Mockito` // `#042` - Comprobaciones de excepciones usando when y thenThrow**
   - _A list of tests with null ids has been added to the class containing the test data._
   - _A test has been created to check that an Exception is thrown when a null is passed to the `findQuestionByExamId` method of the `QuestionRepository` class._